+++ packages/nextra/src/client/components/tabs.tsx
@@ -1,9 +1,15 @@
-import { Tab as HeadlessTab } from '@headlessui/react'
+import {
+  Tab as HeadlessTab,
+  TabGroup,
+  TabList,
+  TabPanel,
+  TabPanels
+} from '@headlessui/react'
 import cn from 'clsx'
 import type { ComponentProps, ReactElement, ReactNode } from 'react'
 import { useCallback, useEffect, useState } from 'react'
 
-type TabItem = string | ReactNode
+type TabItem = string | ReactElement
 
 type TabObjectItem = {
   label: TabItem
@@ -14,7 +20,7 @@ function isTabObjectItem(item: unknown): item is TabObjectItem {
   return !!item && typeof item === 'object' && 'label' in item
 }
 
-function _Tabs({
+function Tabs_({
   items,
   selectedIndex: _selectedIndex,
   defaultIndex = 0,
@@ -75,39 +81,37 @@ function _Tabs({
   }, []) // eslint-disable-line react-hooks/exhaustive-deps -- only on mount
 
   return (
-    <HeadlessTab.Group
+    <TabGroup
       selectedIndex={selectedIndex}
       defaultIndex={defaultIndex}
       onChange={handleChange}
+      tabIndex={-1} // disables focus in Firefox
     >
-      <div className="nextra-scrollbar _overflow-x-auto _overflow-y-hidden _overscroll-x-contain">
-        <HeadlessTab.List className="_mt-4 _flex _w-max _min-w-full _border-b _border-gray-200 _pb-px dark:_border-neutral-800">
-          {items.map((item, index) => {
-            const disabled = isTabObjectItem(item) && item.disabled
-            return (
-              <HeadlessTab
-                key={index}
-                disabled={disabled}
-                className={({ selected }) =>
-                  cn(
-                    '_mr-2 _rounded-t _p-2 _font-medium _leading-5 _transition-colors',
-                    '_-mb-0.5 _select-none _border-b-2',
-                    selected
-                      ? '_border-primary-500 _text-primary-600'
-                      : '_border-transparent _text-gray-600 hover:_border-gray-200 hover:_text-black dark:_text-gray-200 dark:hover:_border-neutral-800 dark:hover:_text-white',
-                    disabled &&
-                      '_pointer-events-none _text-gray-400 dark:_text-neutral-600'
-                  )
-                }
-              >
-                {isTabObjectItem(item) ? item.label : item}
-              </HeadlessTab>
-            )
-          })}
-        </HeadlessTab.List>
-      </div>
-      <HeadlessTab.Panels>{children}</HeadlessTab.Panels>
-    </HeadlessTab.Group>
+      <TabList
+        className={cn(
+          'nextra-scrollbar _overflow-x-auto _overscroll-x-contain _overflow-y-hidden',
+          '_mt-4 _flex _w-full _gap-2 _border-b _border-gray-200 _pb-px dark:_border-neutral-800'
+        )}
+      >
+        {items.map((item, index) => (
+          <HeadlessTab
+            key={index}
+            disabled={isTabObjectItem(item) && item.disabled}
+            className={cn(
+              '_ring-inset',
+              '_rounded-t _p-2 _font-medium _leading-5 _transition-colors',
+              '_-mb-0.5 _select-none _border-b-2',
+              'data-[selected]:!_border-current data-[selected]:!_text-primary-600',
+              '_border-transparent _text-gray-600 hover:_border-gray-200 hover:_text-black dark:_text-gray-200 dark:hover:_border-neutral-800 dark:hover:_text-white',
+              'disabled:_pointer-events-none disabled:_text-gray-400 disabled:dark:_text-neutral-600'
+            )}
+          >
+            {isTabObjectItem(item) ? item.label : item}
+          </HeadlessTab>
+        ))}
+      </TabList>
+      <TabPanels>{children}</TabPanels>
+    </TabGroup>
   )
 }
 
@@ -116,12 +120,12 @@ function Tab({
   // For SEO display all the Panel in the DOM and set `display: none;` for those that are not selected
   unmount = false,
   ...props
-}: Omit<ComponentProps<typeof HeadlessTab.Panel>, 'static'>): ReactElement {
+}: Omit<ComponentProps<typeof TabPanel>, 'static'>): ReactElement {
   return (
-    <HeadlessTab.Panel {...props} unmount={unmount} className="_rounded _pt-6">
+    <TabPanel {...props} unmount={unmount} className="_rounded _mt-6">
       {children}
-    </HeadlessTab.Panel>
+    </TabPanel>
   )
 }
 
-export const Tabs = Object.assign(_Tabs, { displayName: 'Tabs', Tab })
+export const Tabs = Object.assign(Tabs_, { displayName: 'Tabs', Tab })