+++ packages/nextra/src/server/rehype-plugins/rehype.ts
@@ -1,4 +1,4 @@
-import type { Element } from 'hast'
+import type { Element, Root } from 'hast'
 import type { Options as RehypePrettyCodeOptions } from 'rehype-pretty-code'
 import { bundledLanguages, getHighlighter } from 'shiki'
 import type { Plugin } from 'unified'
@@ -27,7 +27,11 @@ export const DEFAULT_REHYPE_PRETTY_CODE_OPTIONS: RehypePrettyCodeOptions = {
     light: 'github-light',
     dark: 'github-dark'
   },
+  defaultLang: {
+    block: 'plaintext'
+  },
   getHighlighter(opts) {
+    // eslint-disable-next-line deprecation/deprecation -- TODO: remove deprecation error
     return getHighlighter({
       ...opts,
       // Without `getHighlighter` option ```mdx lang is not highlighted...
@@ -39,7 +43,7 @@ export const DEFAULT_REHYPE_PRETTY_CODE_OPTIONS: RehypePrettyCodeOptions = {
 
 export const rehypeParseCodeMeta: Plugin<
   [{ defaultShowCopyCode?: boolean }],
-  any
+  Root
 > =
   ({ defaultShowCopyCode }) =>
   ast => {
@@ -66,50 +70,54 @@ export const rehypeParseCodeMeta: Plugin<
     })
   }
 
-export const rehypeAttachCodeMeta: Plugin<[], any> = () => ast => {
-  visit(ast, [{ tagName: 'figure' }, { tagName: 'span' }], (node: Element) => {
-    const isRehypePrettyCode =
-      'data-rehype-pretty-code-figure' in node.properties
-    if (!isRehypePrettyCode) return
+export const rehypeAttachCodeMeta: Plugin<[], Root> = () => ast => {
+  visit(
+    ast as any,
+    [{ tagName: 'figure' }, { tagName: 'span' }],
+    (node: Element) => {
+      const isRehypePrettyCode =
+        'data-rehype-pretty-code-figure' in node.properties
+      if (!isRehypePrettyCode) return
 
-    // remove <figure data-rehype-pretty-code-figure /> element that wraps <pre /> element
-    // because we'll wrap with our own <div />
-    const preEl: PreElement = Object.assign(node, node.children[0])
-    delete preEl.properties['data-theme']
+      // remove <figure data-rehype-pretty-code-figure /> element that wraps <pre /> element
+      // because we'll wrap with our own <div />
+      const preEl: PreElement = Object.assign(node, node.children[0])
+      delete preEl.properties['data-theme']
 
-    if (preEl.tagName === 'pre') {
-      const [codeEl] = preEl.children as Element[]
-      delete codeEl.properties['data-theme']
-      delete codeEl.properties['data-language']
+      if (preEl.tagName === 'pre') {
+        const [codeEl] = preEl.children as Element[]
+        delete codeEl.properties['data-theme']
+        delete codeEl.properties['data-language']
 
-      if (preEl.__hasWordWrap) {
-        preEl.properties['data-word-wrap'] = ''
-      }
+        if (preEl.__hasWordWrap) {
+          preEl.properties['data-word-wrap'] = ''
+        }
 
-      if (preEl.__filename) {
-        preEl.properties['data-filename'] = preEl.__filename
-      }
-      if (preEl.__hasCopyCode) {
-        preEl.properties['data-copy'] = ''
-      }
-      // @ts-expect-error fixme
-      if (preEl.type === 'mdxJsxFlowElement') {
-        if (node.properties.className === undefined)
-          delete node.properties.className
-        if (node.properties.style === undefined) delete node.properties.style
+        if (preEl.__filename) {
+          preEl.properties['data-filename'] = preEl.__filename
+        }
+        if (preEl.__hasCopyCode) {
+          preEl.properties['data-copy'] = ''
+        }
         // @ts-expect-error fixme
-        preEl.attributes.push(
-          ...Object.entries(node.properties).map(([name, value]) => ({
-            type: 'mdxJsxAttribute',
-            name,
-            value: Array.isArray(value) ? value.join(' ') : value
-          }))
-        )
+        if (preEl.type === 'mdxJsxFlowElement') {
+          if (node.properties.className === undefined)
+            delete node.properties.className
+          if (node.properties.style === undefined) delete node.properties.style
+          // @ts-expect-error fixme
+          preEl.attributes.push(
+            ...Object.entries(node.properties).map(([name, value]) => ({
+              type: 'mdxJsxAttribute',
+              name,
+              value: Array.isArray(value) ? value.join(' ') : value
+            }))
+          )
+        }
+      } else {
+        // remove class="line"
+        // @ts-expect-error fixme
+        delete node.children[0].properties.className
       }
-    } else {
-      // remove class="line"
-      // @ts-expect-error fixme
-      delete node.children[0].properties.className
     }
-  })
+  )
 }